EC and CVT unit tests are now proper unit tests that run during the build
