+++ common-tools/clas-analysis/src/main/java/org/jlab/analysis/math/ClasMath.java
@@ -60,7 +60,8 @@ public class ClasMath {
 
 
   public static boolean isWithinXPercent(double X, double val, double standard) {
-    if(val > (1.0 - (X/100.0))*standard && val < (1.0 + (X/100.0))*standard) return true;
+    if(standard >= 0 && val > (1.0 - (X/100.0))*standard && val < (1.0 + (X/100.0))*standard) return true;
+    else if(standard < 0 && val < (1.0 - (X/100.0))*standard && val > (1.0 + (X/100.0))*standard) return true;
     return false;
   }
