+++ reconstruction/cvt/pom.xml
@@ -44,6 +44,18 @@
     </dependency>
 
     <dependency>
+      <groupId>org.jlab.clas</groupId>
+      <artifactId>clas-analysis</artifactId>
+      <version>5.0-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>org.jlab.clas12.detector</groupId>
+      <artifactId>clas12detector-eb</artifactId>
+      <version>1.0-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
       <groupId>cnuphys</groupId>
       <artifactId>swimmer</artifactId>
       <version>2.0-SNAPSHOT</version>
